+++ src/test/groovy/org/m2ci/msp/flaml/FlamlPluginFunctionalTest.groovy
@@ -14,7 +14,8 @@ class FlamlPluginFunctionalTest {
     void setup() {
         def projectDir = File.createTempDir()
         gradle = GradleRunner.create().withPluginClasspath().withProjectDir(projectDir)
-        ['build.gradle', 'foobar.flac', 'foobar.yaml', 'foobar.wav', 'foobar.TextGrid',
+        ['build-core.gradle', 'build-extraction.gradle', 'build-generation.gradle',
+         'foobar.flac', 'foobar.yaml', 'foobar.wav', 'foobar.TextGrid', 'foo_padded.wav', 'bar_padded.wav',
          'foo.wav', 'bar.wav', 'foo.lab', 'bar.lab', 'foo.txt', 'bar.txt'].each { resourceName ->
             new File(projectDir, resourceName).withOutputStream { stream ->
                 stream << this.getClass().getResourceAsStream(resourceName)
@@ -23,7 +24,7 @@ class FlamlPluginFunctionalTest {
     }
 
     @DataProvider
-    Object[][] tasks() {
+    Object[][] coreTasks() {
         [
                 ['help'],
                 ['hasPlugin'],
@@ -32,6 +33,12 @@ class FlamlPluginFunctionalTest {
                 ['hasExtension'],
                 ['hasFlamlResources'],
                 ['hasTestResources'],
+        ]
+    }
+
+    @DataProvider
+    Object[][] extractionTasks() {
+        [
                 ['testExtractWavFiles'],
                 ['testExtractLabFiles'],
                 ['testExtractTextGrid'],
@@ -39,9 +46,33 @@ class FlamlPluginFunctionalTest {
         ]
     }
 
-    @Test(dataProvider = 'tasks')
-    void testTasks(String taskName) {
-        def result = gradle.withArguments(taskName, '-s').build()
+    @DataProvider
+    Object[][] generationTasks() {
+        [
+                ['testGenerateFlac'],
+                ['testGenerateYaml'],
+                ['testInjectText'],
+                ['testInjectSegments']
+        ]
+    }
+
+    @Test(dataProvider = 'coreTasks')
+    void testCoreTasks(String taskName) {
+        def result = gradle.withArguments('--build-file', 'build-core.gradle', taskName).build()
+        println result.output
+        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+    }
+
+    @Test(dataProvider = 'extractionTasks')
+    void testExtractionTasks(String taskName) {
+        def result = gradle.withArguments('--build-file', 'build-extraction.gradle', taskName).build()
+        println result.output
+        assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
+    }
+
+    @Test(dataProvider = 'generationTasks')
+    void testGenerationTasks(String taskName) {
+        def result = gradle.withArguments('--build-file', 'build-generation.gradle', taskName).build()
         println result.output
         assert result.task(":$taskName").outcome in [TaskOutcome.SUCCESS, TaskOutcome.UP_TO_DATE]
     }